Security fix: do not let dbengine grow out of control
There's a race condition in 1.19.0 that can lead to consuming vast
amounts of disk space (I saw 11Gb on our test lab and it was still
growing). Given the little space we have in dom0 and the importance of
not filling the disk entirely to avoid lots of issues, we now fix the
memory mode to 'ram', which means no more disk cache by default with our
package.

See https://github.com/netdata/netdata/issues/8001