Rework patch to fix gcc 4.8 build errors
This patch fixes compiler errors "initializer element is not constant"
in global static structures initialization.

The defines for default struct fields are defined as:

 #define DEFAULT_FOO (struct foo){.bar = 1, ...}

And then used in global static structure declarations as:

 static struct {
     int a;
     int b;
     struct foo f;
     ....;
 } toz = {
     .a = 1,
     .b = 2,
     .f = DEFAULT_FOO,
     ....,
 }

Our gcc 4.8 doesn't like the type cast in DEFAULT_FOO definition,
leading to the compilation error. So this patch removes the typecast in
the default values definition.

But then, in-code assignation fails for lack of typecast:

 struct foo f = DEFAULT_FOO;

Generating the error "expected expression before '{' token".
This patch adds the correct typecast in such cases.

Signed-off-by: Thierry Escande <thierry.escande@vates.tech>