Handle ExecStop in service unit
When removing netdata and netdata-ui packages together (same yum
transaction), the netdata service is restarted by the ui package then
stopped immediately after by the removal the netdata package. This could
make the service to hang and leave it in a failed state and not entirely
removed. This patch adds a delay of 3 seconds before killing the main
netdata process. This give enough time to netdata to start before being
stopped.

Signed-off-by: Thierry Escande <thierry.escande@vates.tech>