Add json-c sources tarball
If the json-c >= 0.17 package is not available, the netdata build system
fetches it from github. Since koji cannot do network fetch and it is not
possible to update json-c in our repos for dependency issues, the json-c
v0.17 tarball is added to the RPM sources. The netdata build system is
patched accordingly to use the sources from the tarball.

Signed-off-by: Thierry Escande <thierry.escande@vates.tech>