Add systemd netdata service stop script
This change adds a script that waits for netdata to be up and running by
waiting for the xenstat plugin to be loaded and executed, then exits or
timeouts after 20 seconds. It exits immediately if netdata is not
running.

This script is used by systemd as ExecStop program to delay the
termination of netdata while it is starting (i.e. calling 'systemctl
start' and 'systemctl stop' immediately after) and avoid the service to
hang in a failed state.

Signed-off-by: Thierry Escande <thierry.escande@vates.tech>